Move reusable components to common folder
