+++ components/common/trading-behavior.tsx
