+++ components/dashboard-components/stats-section.tsx
@@ -1,12 +1,12 @@
-import ChallengesComponent from "./challenges-component";
-import ScallingSection from "./scalling-section";
+import ChallengesOverview from "../common/challenges-overview";
+import ScallingStatus from "../common/scalling-status";
 import UpgradeSection from "./upgrade-section";
 
 export default function StatsSection() {
   return (
     <div className="flex gap-4 justify-between flex-wrap lg:flex-nowrap">
-      <ChallengesComponent />
-      <ScallingSection />
+      <ChallengesOverview />
+      <ScallingStatus />
       <UpgradeSection />
     </div>
   );
